+++ bindings/nodejs/CHANGELOG.md
@@ -19,11 +19,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.1.5 - 2024-01-DD
+
+### Added
+
+- `Utils::{verifyTransactionSyntax(), blockBytes(), blockHashWithoutNonce()}`;
+
 ## 1.1.4 - 2023-12-07
 
 ### Added
 
-- `Utils:verifyTransactionSemantic()`;
+- `Utils::verifyTransactionSemantic()`;
 - `Account::prepareClaimOutputs()` method;
 
 ### Fixed